+++ packages/monaco-editor/index.ts
@@ -1,3 +1,4 @@
+export * from './src/monaco';
 export * from './src/editor';
 export * from './src/utils';
 export * from './src/assist';